move InsertSetAddAssistedColumnsTokenGenerator to rewrite package
